fix: indexDataServiceImpl 삭제 (#13)
