+++ findex/src/main/java/com/example/findex/entity/IndexInfo.java
@@ -18,7 +18,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
 public class IndexInfo extends BaseEntity {
 
   @Column(nullable = false)
